[femutils] Make big methods of BSRFormat and BSRMatrix non inline.
