+++ femutils/FemUtilsGlobal.h
@@ -15,6 +15,7 @@
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
 
+#include "FemDoFsOnNodes.h"
 #include <arcane/utils/ArcaneGlobal.h>
 
 /*---------------------------------------------------------------------------*/
@@ -34,6 +35,7 @@ class CsrRowColumnIndex;
 class IDoFLinearSystemFactory;
 class DoFLinearSystem;
 class IDoFLinearSystemImpl;
+class FemDoFsOnNodes;
 
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
