+++ modules/testlab/CsrGpuBiliAssembly.cc
@@ -14,6 +14,9 @@
 
 #include "FemModule.h"
 
+#include <arcane/accelerator/Scan.h>
+#include <arcane/accelerator/Sort.h>
+
 /*---------------------------------------------------------------------------*/
 /*---------------------------------------------------------------------------*/
